requirements.txt
pytest==3.7.3
requests==2.19.1
pytest-allure-adaptor